refactor: extract budget cards section
