+++ frontend/src/budgets/BudgetsPage.tsx
@@ -1,6 +1,5 @@
-
 import { useCallback, useMemo, useState } from 'react'
-import { AnimatePresence, motion, useReducedMotion } from 'motion/react'
+import { AnimatePresence } from 'motion/react'
 import { Plus } from 'lucide-react'
 import { useSearchParams } from 'react-router-dom'
 import {
@@ -13,7 +12,7 @@ import {
 import { useCategories } from '@/api/categories'
 import { useCurrencies } from '@/api/currency'
 import { useAuth } from '@/hooks/useAuth'
-import BudgetCard from '@/budgets/components/budget-card/BudgetCard'
+import BudgetCardsSection from '@/budgets/components/budget-cards/BudgetCardsSection'
 import BudgetCreateModal from '@/budgets/components/budget-form/BudgetCreateModal'
 import BudgetDetailsModal from '@/budgets/components/budget-details-modal/BudgetDetailsModal'
 import { useBudgetCards } from '@/budgets/hooks/useBudgetCards'
@@ -21,44 +20,11 @@ import { useRecurringBudgetBackfill } from '@/budgets/hooks/useRecurringBudgetBa
 import type { BudgetCardViewModel } from '@/budgets/types'
 import { todayYmd } from '@/budgets/utils/date'
 
-const BUDGET_CARDS_LOADING_AREA_HEIGHT = 'max(0px, calc(100dvh - 15rem))'
-const BUDGET_CARDS_LOADING_TRANSITION = { duration: 0.32, ease: [0.22, 1, 0.36, 1] } as const
-const BUDGET_CARD_ENTER_OFFSET_PX = 12
-const BUDGET_CARD_STAGGER_SECONDS = 0.055
-
-function BudgetCardsLoadingLayer({
-  visible,
-  fill = false,
-}: {
-  visible: boolean
-  fill?: boolean
-}) {
-  const shouldReduceMotion = useReducedMotion() ?? false
-
-  return (
-    <AnimatePresence>
-      {visible && (
-        <motion.div
-          className={[
-            'z-10 flex items-center justify-center overflow-hidden rounded-lg bg-[var(--app-bg)]',
-            fill ? 'absolute inset-0' : 'absolute inset-x-0 top-0',
-          ].join(' ')}
-          style={fill ? undefined : { height: BUDGET_CARDS_LOADING_AREA_HEIGHT }}
-          initial={shouldReduceMotion ? false : { opacity: 0 }}
-          animate={{ opacity: 1 }}
-          exit={{ opacity: 0 }}
-          transition={shouldReduceMotion ? { duration: 0 } : BUDGET_CARDS_LOADING_TRANSITION}
-        >
-          <div className="app-spinner" aria-label="Loading budgets" />
-        </motion.div>
-      )}
-    </AnimatePresence>
-  )
-}
-
+/**
+ * Coordinates budget data loading, URL selection, recurring backfill, and modal workflows
+ */
 export default function BudgetsPage() {
   const { user } = useAuth()
-  const shouldReduceMotion = useReducedMotion() ?? false
   const [searchParams, setSearchParams] = useSearchParams()
   const { data: categories, isLoading: categoriesLoading } = useCategories()
   const { data: currencies, isLoading: currenciesLoading } = useCurrencies()
@@ -143,70 +109,14 @@ export default function BudgetsPage() {
         </button>
       </div>
 
-      {budgetsError && budgetCards.length === 0 ? (
-        <section className="app-card">
-          <p className="text-lg font-semibold" style={{ color: 'var(--app-text)' }}>
-            Budgets could not load
-          </p>
-          <p className="mt-1 text-sm leading-6" style={{ color: 'var(--app-text-subtle)' }}>
-            Refresh the page or try again later.
-          </p>
-        </section>
-      ) : budgetCards.length > 0 ? (
-        <section className="relative" aria-busy={budgetsLoading}>
-          <section className="app-budget-grid">
-            {budgetCards.map((budgetCard, index) => {
-              const { baseBudget, latestPeriod, categoryNames } = budgetCard
-              return (
-                <motion.div
-                  key={baseBudget.id}
-                  className="min-w-0"
-                  initial={shouldReduceMotion ? false : { opacity: 0, y: BUDGET_CARD_ENTER_OFFSET_PX }}
-                  animate={{ opacity: 1, y: 0 }}
-                  transition={{
-                    duration: shouldReduceMotion ? 0 : 0.24,
-                    ease: [0.22, 1, 0.36, 1],
-                    delay: shouldReduceMotion ? 0 : index * BUDGET_CARD_STAGGER_SECONDS,
-                  }}
-                >
-                  <BudgetCard
-                    baseBudget={baseBudget}
-                    latestPeriod={latestPeriod}
-                    categoryNames={categoryNames}
-                    utilization={latestPeriod ? latestUtilizationByBudgetId.get(latestPeriod.id) : undefined}
-                    onOpen={() => openBudget(budgetCard)}
-                  />
-                </motion.div>
-              )
-            })}
-          </section>
-          <BudgetCardsLoadingLayer visible={budgetsLoading} />
-        </section>
-      ) : budgetsLoading ? (
-        <section
-          className="relative overflow-hidden rounded-lg"
-          style={{ height: BUDGET_CARDS_LOADING_AREA_HEIGHT }}
-          aria-busy
-        >
-          <BudgetCardsLoadingLayer visible fill />
-        </section>
-      ) : (
-        <section className="flex min-h-[calc(100vh-16rem)] items-center justify-center text-center italic text-sm" style={{ color: 'var(--app-text-subtle)' }}>
-          <div>
-            <p>
-              No budgets to display
-            </p>
-            <p className="mt-1">
-              Create a budget to start tracking limits, spending, and category progress.
-            </p>
-          </div>
-          {(categoriesLoading || currenciesLoading) && (
-            <p className="sr-only">
-              Loading form options...
-            </p>
-          )}
-        </section>
-      )}
+      <BudgetCardsSection
+        budgetCards={budgetCards}
+        latestUtilizationByBudgetId={latestUtilizationByBudgetId}
+        loading={budgetsLoading}
+        error={budgetsError}
+        formOptionsLoading={categoriesLoading || currenciesLoading}
+        onOpenBudget={openBudget}
+      />
 
       <BudgetCreateModal
         key={`${defaultCurrency}-${userTimeZone}`}
